Move polka linking to compilation phase, add polkavm-wrapper to invoke compiled modules
