+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "language/move-native",
     "language/polkavm/move-to-polka",
     "language/tools/move-mv-llvm-compiler",
+    "language/tools/polkavm-wrapper",
 ]
 
 # NOTE: default-members is the complete list of binaries that form the "production Move codebase". These members should
@@ -12,11 +13,16 @@ members = [
 # test-only code included.
 #
 # For more, see the "Conditional compilation for tests" section in documentation/coding_guidelines.md.
-default-members = ["language/move-native", "language/polkavm/move-to-polka"]
+default-members = [
+    "language/move-native",
+    "language/polkavm/move-to-polka",
+    "language/tools/polkavm-wrapper",
+]
 
 # Dependencies that should be kept in sync through the whole workspace
 [workspace.dependencies]
 bcs = "0.1.4"
+
 
 [profile.bench]
 debug = true
