+++ language/polkavm/move-to-polka/Cargo.toml
@@ -38,6 +38,7 @@ num = "0.4.0"
 num-traits = "0.2"
 once_cell = "1.21"
 parking_lot = "0.12"
+polkavm-linker = "0.22.0"
 regex = "1.1.9"
 semver = "1.0.13"
 serde = { version = "1.0.124", features = ["derive"] }
@@ -48,7 +49,6 @@ which = "7.0"
 
 [dev-dependencies]
 polkavm = "0.22.0"
-polkavm-linker = "0.22.0"
 
 [build-dependencies]
 move-stdlib = { git = "https://github.com/move-language/move-on-aptos.git", package = "move-stdlib" }
