+++ language/polkavm/move-to-polka/src/linker.rs
@@ -1,0 +1,9 @@
+pub fn load_from_elf_with_polka_linker(data: &[u8]) -> anyhow::Result<Vec<u8>> {
+    // config is taken from polkatool with default values
+    let mut config = polkavm_linker::Config::default();
+    config.set_strip(false);
+    config.set_optimize(true);
+
+    let res = polkavm_linker::program_from_elf(config, data)?;
+    Ok(res)
+}
