+++ language/tools/polkavm-wrapper/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "polkavm-wrapper"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1.0.98"
+clap = { version = "4.5.37", features = ["derive"] }
+polkavm = "0.22.0"
+tracing = "0.1.41"
+tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
